+++ test/windows/test_shellidlist.h
@@ -3,6 +3,7 @@
 
 #include "../core/test.h"
 #include <qewindows/idlist.h>
+#include <qewindows/shell.h>
 #include <qdebug.h>
 
 qe::windows::shell::IdList copy_from_path(LPCWSTR path)
@@ -16,9 +17,9 @@ qe::windows::shell::IdList copy_from_path(LPCWSTR path)
 void dump_id(qe::windows::shell::IdList &id)
 {
     qDebug() << "Address: " << id.data()
-             << " Type:    "<< id.rawType()
-             << " Length:  " << id.elementCount()
-             << " Bytes:   " << id.byteCount() << '\n';
+             << "Type:    "<< id.rawType()
+             << "Length:  " << id.elementCount()
+             << "Bytes:   " << id.byteCount();
 }
 
 struct test_shellidlist
@@ -26,8 +27,9 @@ struct test_shellidlist
     static void run()
     {
         CoInitialize(nullptr);
-        test_construction();
-        test_iteration();
+        test_shellidlist::test_construction();
+        test_shellidlist::test_iteration();
+        test_shellidlist::test_accessors();
         CoUninitialize();
     }
 
@@ -66,6 +68,35 @@ struct test_shellidlist
         for (const auto &iter3 : id1) {
             (void)iter3;
         }
+    }
+
+    static void test_accessors()
+    {
+        using namespace qe::windows::shell;
+        IdList id1(copy_from_path(L"C:\\Program Files"));
+        EXPECT_TRUE(id1);
+
+        //check that the byte count makes sense
+        EXPECT_GT(id1.byteCount(), 5u);
+        //This PC -> C: -> Program Files
+        EXPECT_EQ(id1.elementCount(), 3);
+        EXPECT_FLAG(unsigned int, id1.inferredType(), InferredType::Folder);
+
+        IdList id2 = id1.parent();
+        EXPECT_TRUE(id2);
+        EXPECT_NE(id2, id1);
+        EXPECT_TRUE(ILIsParent(id2.castTo<PCIDLIST_ABSOLUTE>(),
+                               id1.castTo<PCIDLIST_ABSOLUTE>(), TRUE));
+
+        EXPECT_EQ(id2.elementCount(), 2);
+        EXPECT_FLAG(unsigned int, id2.inferredType(), InferredType::Drive);
+
+        IdList id3 = id2.parent();
+        EXPECT_TRUE(id3);
+        EXPECT_NE(id3, id2);
+        EXPECT_NE(id3, id1);
+
+        EXPECT_FLAG(unsigned int, id3.inferredType(), InferredType::LocalVirtual);
     }
 };